+++ jafar_sketch/rx5808.ino
@@ -22,7 +22,7 @@
 #include "rx5808.h"
 #define RSSI_THRESH (scanVec[getMaxPos()]-5) //strong channels are near to the global max
 
-RX5808::RX5808(uint16_t RSSIpin, uint16_t CSpin) {
+RX5808::RX5808(uint8_t RSSIpin, uint8_t CSpin) {
   _rssiPin = RSSIpin;
   _csPin = CSpin;
   _stop_scan = 0;
